requirements.txt
fonttools==4.34.4
hydra-core==1.2.0
icecream==2.1.3
matplotlib==3.5.2
numpy==1.23.1
omegaconf~=2.2.2
pandas==1.4.3
scipy==1.8.1
seaborn==0.11.2
tabulate==0.8.10
tensorboard==2.9.1
torch==1.12.0
tqdm==4.64.0
